+++ cli/src/config.rs
@@ -12,6 +12,12 @@ use std::{
 pub struct ReinferConfig {
     current_context: Option<String>,
     contexts: Vec<ContextConfig>,
+    #[serde(default = "default_true")]
+    pub context_is_required: bool,
+}
+
+fn default_true() -> bool {
+    true
 }
 
 impl ReinferConfig {