+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+- BREAKING: the `--context` option is now required. Users need to opt
+  out if they don't want to provide this for every command
+- BREAKING: the `--context` option is always a required field for internal users
+
 # v0.23.0
 
 - Add `get emails`